重构代码

+++ src/main/java/com/meread/selenium/service/WebDriverManagerLocal.java
@@ -167,7 +167,7 @@ public class WebDriverManagerLocal extends BaseWebDriverManager {
     }
 
     @Override
-    public void releaseWebDriver(String removeChromeSessionId) {
+    public void releaseWebDriver(String removeChromeSessionId,boolean quit) {
         Iterator<Map.Entry<String, MyChrome>> iterator = chromes.entrySet().iterator();
         while (iterator.hasNext()) {
             MyChrome myChrome = iterator.next().getValue();
@@ -184,7 +184,7 @@ public class WebDriverManagerLocal extends BaseWebDriverManager {
                         clientExpireTime = client.getExpireTime();
                     }
                     //chrome的存活时间不够一个opTime时间，则chrome不退出，只清理客户端引用
-                    if ((chromeExpireTime - clientExpireTime) / 1000 > opTimeout) {
+                    if ((chromeExpireTime - clientExpireTime) / 1000 > opTimeout && !quit) {
                         myChrome.setUserTrackId(null);
                         clients.remove(userTrackId);
                         myChrome.getWebDriver().manage().deleteAllCookies();
